Update some tests
