+++ lib/commands/context/helpers.js
@@ -55,7 +55,7 @@ export const NATIVE_WIN = 'NATIVE_APP';
 export const WEBVIEW_WIN = 'WEBVIEW';
 export const CHROMIUM_WIN = 'CHROMIUM';
 export const WEBVIEW_BASE = `${WEBVIEW_WIN}_`;
-const DEVTOOLS_SOCKET_PATTERN = /@[\w.]+_devtools_remote_?([\w.]+_)?(\d+)?\b/;
+export const DEVTOOLS_SOCKET_PATTERN = /@[\w.]+_devtools_remote_?([\w.]+_)?(\d+)?\b/;
 const WEBVIEW_PID_PATTERN = new RegExp(`^${WEBVIEW_BASE}(\\d+)`);
 const WEBVIEW_PKG_PATTERN = new RegExp(`^${WEBVIEW_BASE}([^\\d\\s][\\w.]*)`);
 const WEBVIEW_WAIT_INTERVAL_MS = 200;
