+++ lib/commands/performance.js
@@ -3,7 +3,7 @@ import {retryInterval} from 'asyncbox';
 import _ from 'lodash';
 import {requireArgs} from '../utils';
 
-const NETWORK_KEYS = [
+export const NETWORK_KEYS = [
   [
     'bucketStart',
     'activeTime',
@@ -16,9 +16,9 @@ const NETWORK_KEYS = [
   ],
   ['st', 'activeTime', 'rb', 'rp', 'tb', 'tp', 'op', 'bucketDuration'],
 ];
-const CPU_KEYS = /** @type {const} */ (['user', 'kernel']);
-const BATTERY_KEYS = ['power'];
-const MEMORY_KEYS = [
+export const CPU_KEYS = /** @type {const} */ (['user', 'kernel']);
+export const BATTERY_KEYS = ['power'];
+export const MEMORY_KEYS = [
   'totalPrivateDirty',
   'nativePrivateDirty',
   'dalvikPrivateDirty',
@@ -35,7 +35,7 @@ const MEMORY_KEYS = [
   'dalvikRss',
   'totalRss',
 ];
-const SUPPORTED_PERFORMANCE_DATA_TYPES = Object.freeze({
+export const SUPPORTED_PERFORMANCE_DATA_TYPES = Object.freeze({
   cpuinfo:
     'the amount of cpu by user and kernel process - cpu information for applications on real devices and simulators',
   memoryinfo:
@@ -45,7 +45,7 @@ const SUPPORTED_PERFORMANCE_DATA_TYPES = Object.freeze({
   networkinfo:
     'the network statistics - network rx/tx information for applications on real devices and simulators',
 });
-const MEMINFO_TITLES = Object.freeze({
+export const MEMINFO_TITLES = Object.freeze({
   NATIVE: 'Native',
   DALVIK: 'Dalvik',
   EGL: 'EGL',
@@ -225,7 +225,7 @@ function parseMeminfoForApiAbove29(entries, valDict) {
  * @param {string} packageName
  * @param {number} retries
  */
-async function getMemoryInfo(packageName, retries = 2) {
+export async function getMemoryInfo(packageName, retries = 2) {
   return await retryInterval(retries, RETRY_PAUSE_MS, async () => {
     const cmd = [
       'dumpsys',
@@ -267,7 +267,7 @@ async function getMemoryInfo(packageName, retries = 2) {
  * @this {AndroidDriver}
  * @param {number} retries
  */
-async function getNetworkTrafficInfo(retries = 2) {
+export async function getNetworkTrafficInfo(retries = 2) {
   return await retryInterval(retries, RETRY_PAUSE_MS, async () => {
     let returnValue = [];
     let bucketDuration, bucketStart, activeTime, rxBytes, rxPackets, txBytes, txPackets, operations;
@@ -450,7 +450,7 @@ async function getNetworkTrafficInfo(retries = 2) {
  *                  '14.3' is usage by the user (%), '28.2' is usage by the kernel (%)
  * @throws {Error} If it failed to parse the result of dumpsys, or no package name exists.
  */
-async function getCPUInfo(packageName, retries = 2) {
+export async function getCPUInfo(packageName, retries = 2) {
   // TODO: figure out why this is
   // sometimes, the function of 'adb.shell' fails. when I tested this function on the target of 'Galaxy Note5',
   // adb.shell(dumpsys cpuinfo) returns cpu datas for other application packages, but I can't find the data for packageName.
@@ -494,7 +494,7 @@ async function getCPUInfo(packageName, retries = 2) {
  * @this {AndroidDriver}
  * @param {number} retries
  */
-async function getBatteryInfo(retries = 2) {
+export async function getBatteryInfo(retries = 2) {
   return await retryInterval(retries, RETRY_PAUSE_MS, async () => {
     let cmd = ['dumpsys', 'battery', '|', 'grep', 'level'];
     let data = await this.adb.shell(cmd);
