+++ lib/commands/system-bars.js
@@ -33,7 +33,7 @@ export async function getSystemBars() {
       `Cannot retrieve system bars details. Original error: ${/** @type {Error} */ (e).message}`,
     );
   }
-  return parseWindows(stdout, this.log);
+  return parseWindows.bind(this)(stdout);
 }
 
 /**
@@ -86,19 +86,19 @@ export async function mobilePerformStatusBarCommand(opts) {
 /**
  * Parses window properties from adb dumpsys output
  *
+ * @this {import('../driver').AndroidDriver}
  * @param {string} name The name of the window whose properties are being parsed
  * @param {Array<string>} props The list of particular window property lines.
  * Check the corresponding unit tests for more details on the input format.
- * @param {import('@appium/types').AppiumLogger} [log] Logger instance
  * @returns {WindowProperties} Parsed properties object
  * @throws {Error} If there was an issue while parsing the properties string
  */
-export function parseWindowProperties(name, props, log) {
+export function parseWindowProperties(name, props) {
   const result = _.cloneDeep(DEFAULT_WINDOW_PROPERTIES);
   const propLines = props.join('\n');
   const frameMatch = FRAME_PATTERN.exec(propLines);
   if (!frameMatch) {
-    log?.debug(propLines);
+    this.log.debug(propLines);
     throw new Error(`Cannot parse the frame size from '${name}' window properties`);
   }
   result.x = parseFloat(frameMatch[1]);
@@ -107,7 +107,7 @@ export function parseWindowProperties(name, props, log) {
   result.height = parseFloat(frameMatch[4]) - result.y;
   const visibilityMatch = VIEW_VISIBILITY_PATTERN.exec(propLines);
   if (!visibilityMatch) {
-    log?.debug(propLines);
+    this.log.debug(propLines);
     throw new Error(`Cannot parse the visibility value from '${name}' window properties`);
   }
   result.visible = parseInt(visibilityMatch[1], 16) === VIEW_VISIBLE;
@@ -117,14 +117,14 @@ export function parseWindowProperties(name, props, log) {
 /**
  * Extracts status and navigation bar information from the window manager output.
  *
+ * @this {import('../driver').AndroidDriver}
  * @param {string} lines Output from dumpsys command.
  * Check the corresponding unit tests for more details on the input format.
- * @param {import('@appium/types').AppiumLogger} [log] Logger instance
  * @return {StringRecord} An object containing two items where keys are statusBar and navigationBar,
  * and values are corresponding WindowProperties objects
  * @throws {Error} If no window properties could be parsed
  */
-export function parseWindows(lines, log) {
+export function parseWindows(lines) {
   /**
    * @type {StringRecord}
    */
@@ -147,7 +147,7 @@ export function parseWindows(lines, log) {
     }
   }
   if (_.isEmpty(windows)) {
-    log?.debug(lines);
+    this.log.debug(lines);
     throw new Error('Cannot parse any window information from the dumpsys output');
   }
 
@@ -155,9 +155,9 @@ export function parseWindows(lines, log) {
   const result = {};
   for (const [name, props] of _.toPairs(windows)) {
     if (name.startsWith(STATUS_BAR_WINDOW_NAME_PREFIX)) {
-      result.statusBar = parseWindowProperties(name, props, log);
+      result.statusBar = parseWindowProperties.bind(this)(name, props);
     } else if (name.startsWith(NAVIGATION_BAR_WINDOW_NAME_PREFIX)) {
-      result.navigationBar = parseWindowProperties(name, props, log);
+      result.navigationBar = parseWindowProperties.bind(this)(name, props);
     }
   }
   const unmatchedWindows = /** @type {const} */ ([
@@ -165,7 +165,7 @@ export function parseWindows(lines, log) {
     ['navigationBar', NAVIGATION_BAR_WINDOW_NAME_PREFIX],
   ]).filter(([name]) => _.isNil(result[name]));
   for (const [window, namePrefix] of unmatchedWindows) {
-    log?.info(
+    this.log.info(
       `No windows have been found whose title matches to ` +
         `'${namePrefix}'. Assuming it is invisible. ` +
         `Only the following windows are available: ${_.keys(windows)}`,
